Profile results records

+++ complete/src/edument/perl6idea/profiler/run/Perl6ProfileTask.java
@@ -4,7 +4,6 @@ import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationType;
 import com.intellij.notification.Notifications;
 import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.progress.ProcessCanceledException;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.Task;
@@ -18,21 +17,31 @@ import javax.swing.*;
 import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 public class Perl6ProfileTask extends Task.Backgroundable {
-    public static final Logger LOG = Logger.getInstance(Perl6ProfileTask.class);
     private final ProfilerView myProfilerView;
     private File sqlDataFile;
     private Perl6ProfileData myProfileData;
+    private boolean myHasToRemoveTheFile;
 
     public Perl6ProfileTask(Project project,
-                            String data,
+                            String title,
                             boolean canBeCancelled,
                             File file,
-                            ProfilerView profilerView) {
-        super(project, data, canBeCancelled);
+                            ProfilerView profilerView, boolean hasToRemoveTheFile) {
+        super(project, title, canBeCancelled);
         sqlDataFile = file;
         myProfilerView = profilerView;
+        myHasToRemoveTheFile = hasToRemoveTheFile;
+    }
+
+    public Perl6ProfileTask(Project project, String title, boolean canBeCancelled, Perl6ProfileData data, ProfilerView view) {
+        super(project, title, canBeCancelled);
+        myProfileData = data;
+        myProfilerView = view;
     }
 
     @Override
@@ -42,10 +51,16 @@ public class Perl6ProfileTask extends Task.Backgroundable {
             myProfileData.cancel();
             myProfileData = null;
         }
-        if (sqlDataFile != null) {
+        if (sqlDataFile != null && myHasToRemoveTheFile) {
             sqlDataFile.delete();
             sqlDataFile = null;
         }
+    }
+
+    private static String createProfileName() {
+        DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
+        Date date = new Date();
+        return "Profile results at " + dateFormat.format(date);
     }
 
     @Override
@@ -56,7 +71,7 @@ public class Perl6ProfileTask extends Task.Backgroundable {
             indicator.setText("Creating a database...");
             indicator.setFraction(0.1);
             indicator.checkCanceled();
-            myProfileData = new Perl6ProfileData(sqlDataFile.getCanonicalPath());
+            myProfileData = myProfileData == null ? new Perl6ProfileData(myProject, createProfileName(), sqlDataFile) : myProfileData;
             indicator.setText("Loading profiler data into the database...");
             indicator.setFraction(0.2);
             myProfileData.initialize();
@@ -75,7 +90,7 @@ public class Perl6ProfileTask extends Task.Backgroundable {
             myProfilerView.setView(new JLabel("Could not collect profiling results"));
             throw new ProcessCanceledException();
         } finally {
-            if (sqlDataFile != null) {
+            if (sqlDataFile != null && myHasToRemoveTheFile) {
                 sqlDataFile.delete();
                 sqlDataFile = null;
             }
